refactor(extractor): promote KVS-generic helpers from abxxx/ to base/kvs/
Three pieces of the abxxx extractor were stock-KVS, not site-specific,
and would be re-implemented verbatim by every sibling deployment in the
"OUR NETWORK" footer (AbJav, AbGay, AbTranny, AbMilf, AbBDSM, AbEbony,
AbLesbian, …). Promoting them now closes the most expensive part of any
follow-up extractor sprint to a one-liner URL prefix change.

Layout

    crates/rdlp-extractor/src/base/kvs.rs
      → crates/rdlp-extractor/src/base/kvs/
          ├── mod.rs               (re-exports flashvars surface; declares submodules)
          ├── flashvars.rs         (existing flashvars parser, unchanged)
          ├── url_obfuscation.rs   (was abxxx/decode.rs)
          ├── api.rs               (lookup endpoint + KvsVideoMetadata + parser)
          └── file_formats.rs      (parse_file_formats_blob + parse_search_for_file_formats)

What moved (and what stayed)

KVS-generic, promoted:
- url_obfuscation::decode_video_url        — Cyrillic homoglyph + comma-split base64
- api::video_lookup_endpoint{,_with_lifetime} — id-bucketed JSON lookup URL
- api::KvsVideoMetadata + parse_video_lookup  — canonical {video:{…}} schema
- file_formats::parse_file_formats_blob       — packed ||{ext}|{WxH}|… parser
- file_formats::parse_search_for_file_formats — find row by id, project blob+dims

Site-specific, stayed in abxxx/:
- patterns.rs                — abxxx.com URL regex
- metadata::slug_search_endpoint — sort/section choice may differ across siblings
- mod.rs::extract            — wiring (which endpoints, in which order, fallbacks)

API surface is pub(crate) throughout — the existing visibility pattern of
base/* — so no external callers are affected. base::kvs's pre-existing
flashvars surface (KvsFlashvars, parse_kvs_flashvars, KvsFormat,
extract_kvs_formats, KvsMetadata, extract_kvs_metadata, is_kvs_page) is
re-exported from the new mod.rs so XTits and Generic still compile
without import-path changes.

Verification

- 874 / 874 extractor lib tests green (48 abxxx + kvs)
- cargo clippy --workspace --all-targets -- -D warnings clean
- Live extraction byte-identical with the pre-refactor result:
  width=1280, height=720, filesize=853473694, format_note=720p,
  uploader=Xavier Torres, view_count=87637

abxxx/decode.rs is removed (git mv → base/kvs/url_obfuscation.rs).
abxxx/metadata.rs shrinks to just slug_search_endpoint (~30 lines).
